Blindly fetch all submodules in regressions
If you have a situation where a user has ROCKETCHIP_ADDONS that are
not submodules, but are included as explicit subdirectories in a
branch, the regressions will fail trying to checkout a non-existent
submodule. Since riscv-tools/rocket-tools is now NOT a submodule, it
should be safe to do an explicit clone of all submodules without
filtering to the hardware-specific ones.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>